Rating_System_Console_App
A Program that shows topics and you rate them 1-10.
After one user goes through them you can have another user rate them or end the program.
At the end it will print the results,
Show the topic with the highest points and the topic with the lowest points. The average will also be printed for each topic.